+++ asdf/config.py
@@ -83,16 +83,26 @@ class AsdfConfig:
         mapping : collections.abc.Mapping, optional
             Mapping to remove.
         package : str, optional
-            Python package whose mappings will all be removed.
+            Remove only extensions provided by this package.  If the `mapping`
+            argument is omitted, then all mappings from this package will
+            be removed.
         """
-        with self._lock:
-            resource_mappings = self.resource_mappings
-            if mapping is not None:
+        if mapping is None and package is None:
+            raise ValueError("Must specify at least one of mapping or package")
+
+        if mapping is not None:
                 mapping = ResourceMappingProxy.maybe_wrap(mapping)
-                resource_mappings = [m for m in resource_mappings if m != mapping]
+
+        def _remove_condition(m):
+            result = True
+            if mapping is not None:
+                result = result and m == mapping
             if package is not None:
-                resource_mappings = [m for m in resource_mappings if m.package_name != package]
-            self._resource_mappings = resource_mappings
+                result = result and m.package_name == package
+            return result
+
+        with self._lock:
+            self._resource_mappings = [m for m in self.resource_mappings if not _remove_condition(m)]
             self._resource_manager = None
 
     def reset_resources(self):
@@ -154,19 +164,58 @@ class AsdfConfig:
 
         Parameters
         ----------
-        extension : asdf.extension.AsdfExtension, optional
-            An extension instance to remove.
+        extension : asdf.extension.AsdfExtension or str, optional
+            An extension instance or URI to remove.
         package : str, optional
-            A Python package name whose extensions will all be removed.
+            Remove only extensions provided by this package.  If the `extension`
+            argument is omitted, then all extensions from this package will
+            be removed.
         """
-        with self._lock:
-            extensions = self.extensions
-            if extension is not None:
-                extension = ExtensionProxy.maybe_wrap(extension)
-                extensions = [e for e in extensions if e != extension]
+        if extension is None and package is None:
+            raise ValueError("Must specify at least one of extension or package")
+
+        if extension is not None and not isinstance(extension, str):
+            extension = ExtensionProxy.maybe_wrap(extension)
+
+        def _remove_condition(e):
+            result = True
+
+            if isinstance(extension, str):
+                result = result and e.extension_uri == extension
+            elif isinstance(extension, ExtensionProxy):
+                result = result and e == extension
+
             if package is not None:
-                extensions = [e for e in extensions if e.package_name != package]
-            self._extensions = extensions
+                result = result and e.package_name == package
+
+            return result
+
+        with self._lock:
+            self._extensions = [e for e in self.extensions if not _remove_condition(e)]
+
+    def get_extension(self, extension_uri):
+        """
+        Get an extension by URI.
+
+        Parameters
+        ----------
+        extension_uri : str
+            The URI of the extension to fetch.
+
+        Returns
+        -------
+        asdf.extension.AsdfExtension
+
+        Raises
+        ------
+        KeyError
+            If no such extension exists.
+        """
+        for extension in self.extensions:
+            if extension.extension_uri == extension_uri:
+                return extension
+
+        raise KeyError("No registered extension with URI '{}'".format(extension_uri))
 
     def reset_extensions(self):
         """