+++ CHANGES.rst
@@ -16,6 +16,9 @@
 
 - Drop support for Python 3.5. [#856]
 
+- Add new extension API to support versioned extensions.
+  [#850]
+
 2.7.0 (2020-07-23)
 ------------------